+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.3.0
+
+##### Breaking
+- Updates the native iOS bridge to use [Appboy iOS SDK 2.29.0](https://github.com/Appboy/appboy-ios-sdk/blob/master/CHANGELOG.md#2290), which drops support for iOS 7.
+- Updates the native Android bridge to use [Appboy Android SDK 2.0.0](https://github.com/Appboy/appboy-android-sdk/blob/master/CHANGELOG.md#200).
+
+##### Added
+- Adds `ReactAppboy.requestImmediateDataFlush()` for requesting an immediate flush of any data waiting to be sent to Appboy's servers.
+- Adds `ReactAppboy.requestFeedRefresh()` for requesting a refresh of the News Feed.
+  - See https://github.com/Appboy/appboy-react-sdk/pull/12. Thanks @stief510!
+- Added the ability to pass an optional dictionary of News Feed launch options to `launchNewsFeed()`. See `NewsFeedLaunchOptions` for supported keys.
+  - For more information on currently supported `NewsFeedLaunchOptions` keys, see the card width and card margin properties on [ABKFeedViewController](http://appboy.github.io/appboy-ios-sdk/docs/interface_a_b_k_feed_view_controller.html).
+  - See https://github.com/Appboy/appboy-react-sdk/pull/10. Thanks @mihalychk!
+
 ## 1.2.0
 
 ##### Breaking